Revert "Add authorization request cancel"
This reverts commit 211f6b15329ea64c333f302927faaa1f4f13ac10.
