+++ sdk/src/test/java/com/iovation/launchkey/sdk/client/BasicServiceClientTestGetAuthorizationResponse.java
@@ -26,7 +26,7 @@ import static org.mockito.Mockito.when;
 
 
 @RunWith(Enclosed.class)
-public class BasicServiceClientGetAuthorizationResponseTest extends TestCase {
+public class BasicServiceClientTestGetAuthorizationResponse extends TestCase {
 
     @RunWith(MockitoJUnitRunner.class)
     public static class StandardTests {
