+++ sdk/src/test/java/com/iovation/launchkey/sdk/client/BasicServiceClientTestSessionEnd.java
@@ -2,6 +2,7 @@ package com.iovation.launchkey.sdk.client;
 
 import com.iovation.launchkey.sdk.transport.Transport;
 import com.iovation.launchkey.sdk.transport.domain.EntityIdentifier;
+import com.iovation.launchkey.sdk.transport.domain.ServiceV3SessionsDeleteRequest;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -14,17 +15,20 @@ import java.util.UUID;
 
 import static org.junit.Assert.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.verify;
 
 @RunWith(MockitoJUnitRunner.class)
-public class BasicServiceClientCancelAuthorizationRequestTest {
+public class BasicServiceClientTestSessionEnd {
     private final static UUID serviceId = UUID.fromString("67c87654-aed9-11e7-98e9-0469f8dc10a5");
-    private final static String authRequestId = "a5a07697-92ab-4c45-81d6-1ea41c019c09";
+    private final static String user = "User Name";
     @Mock
     public Transport transport;
 
     @Captor
     private ArgumentCaptor<EntityIdentifier> entityCaptor;
+
+    @Captor
+    private ArgumentCaptor<ServiceV3SessionsDeleteRequest> requestCaptor;
 
     private ServiceClient client;
     @Before
@@ -32,26 +36,24 @@ public class BasicServiceClientCancelAuthorizationRequestTest {
         client = new BasicServiceClient(serviceId, transport);
     }
 
-
-    @Test
-    public void sendsAuthorizationRequestId() throws Exception {
-        client.cancelAuthorizationRequest(authRequestId);
-        UUID expected = UUID.fromString("a5a07697-92ab-4c45-81d6-1ea41c019c09");
-        verify(transport).serviceV3AuthsDelete(eq(expected), any(EntityIdentifier.class));
-    }
-
-
     @Test
     public void sendsSubjectEntityType() throws Exception {
-        client.cancelAuthorizationRequest(authRequestId);
-        verify(transport).serviceV3AuthsDelete(any(UUID.class), entityCaptor.capture());
+        client.sessionEnd(user);
+        verify(transport).serviceV3SessionsDelete(any(ServiceV3SessionsDeleteRequest.class), entityCaptor.capture());
         assertEquals(EntityIdentifier.EntityType.SERVICE, entityCaptor.getValue().getType());
     }
 
     @Test
     public void sendsSubjectEntityId() throws Exception {
-        client.cancelAuthorizationRequest(authRequestId);
-        verify(transport).serviceV3AuthsDelete(any(UUID.class), entityCaptor.capture());
+        client.sessionEnd(user);
+        verify(transport).serviceV3SessionsDelete(any(ServiceV3SessionsDeleteRequest.class), entityCaptor.capture());
         assertEquals(serviceId, entityCaptor.getValue().getId());
+    }
+
+    @Test
+    public void sendsUserAsEndUserIdentifier() throws Exception {
+        client.sessionEnd(user);
+        verify(transport).serviceV3SessionsDelete(requestCaptor.capture(), any(EntityIdentifier.class));
+        assertEquals(user, requestCaptor.getValue().getEndUserIdentifier());
     }
 }
